Implemented hybi17 on top of 10
