+++ Handlers/WebSocket/hybi10/DataFrame.cs
@@ -31,7 +31,7 @@ namespace Alchemy.Server.Handlers.WebSocket.hybi10
     /// Automatically manages adding received data to an existing frame and checking whether or not we've received the entire frame yet.
     /// See http://www.whatwg.org/specs/web-socket-protocol/ for more details on the WebSocket Protocol.
     /// </summary>
-    public class DataFrame : Alchemy.Server.Classes.DataFrame
+    public class DataFrame : Alchemy.Server.Handlers.WebSocket.DataFrame
     {
         public enum OpCode
         {
